+++ .changeset/kcneu3k8.md
@@ -1,0 +1,5 @@
+---
+'@bbodek/internal-ui': patch
+---
+
+피드백 반영
