+++ apps/internal-ui/src/components/Badge/Badge/index.ts
@@ -1,0 +1,3 @@
+export { default } from './Badge';
+export * from './constants';
+export * from './types';
