+++ apps/internal-ui/src/components/Badge/InfoBadge/index.ts
@@ -1,0 +1,3 @@
+export { default } from './InfoBadge';
+export * from './constants';
+export * from './types';
