+++ apps/internal-ui/src/components/Badge/index.ts
@@ -1,4 +1,8 @@
 export { default as Badge } from './Badge';
+export * from './Badge/constants';
+export * from './Badge/types';
 export { default as InfoBadge } from './InfoBadge';
-export * from './constants';
-export * from './types';
+export * from './InfoBadge/constants';
+export * from './InfoBadge/types';
+export * from './shared/constants';
+export * from './shared/types';
