+++ apps/internal-ui/src/components/Badge/shared/constants/index.ts
@@ -1,0 +1,7 @@
+export const BADGE_THEMES = {
+  GRAY: 'gray',
+  PRIMARY: 'primary',
+  GREEN: 'green',
+  YELLOW: 'yellow',
+  RED: 'red',
+} as const;
