+++ apps/internal-ui/src/components/Badge/shared/index.ts
@@ -1,0 +1,2 @@
+export * from './constants';
+export * from './types';
