+++ apps/internal-ui/src/components/Badge/shared/types/index.ts
@@ -1,0 +1,10 @@
+import { BADGE_THEMES } from '@/components/Badge/shared/constants';
+
+export type BadgeTheme = (typeof BADGE_THEMES)[keyof typeof BADGE_THEMES];
+
+export interface BadgeThemeStyle {
+  CONTAINER: string;
+  LABEL: string;
+}
+
+export type BadgeStyle = Record<BadgeTheme, BadgeThemeStyle>;
